Add files via upload
The system that is identified, that helps to apply leave through online to the higher authorities which in return sents an SMTP mail message to the applicant regarding the status  the leave. Leave management system makes use of the features of the UiPath to maintain the leave records of the staffs and students updated each and every time. Admin makes some norms that are set for both students and staffs as per the guidelines of the organization. The system would help to make easier maintenance ofrecordsanddecreasethepaperwork.